Refactor redundant integration test utility methods
