+++ service/src/integration-test/groovy/org/olf/ILLBrokerSpec.groovy
@@ -29,9 +29,11 @@ class ILLBrokerSpec extends TestBase {
 
     final static String TENANT_ONE_NAME = "BrokerInstOne"
     final static String TENANT_TWO_NAME = "BrokerInstTwo"
+    final static String TENANT_THREE_NAME = "BrokerInstThree"
     final static String SYMBOL_AUTHORITY = "ISIL"
     final static String SYMBOL_ONE_NAME = "BIO1"
     final static String SYMBOL_TWO_NAME = "BIO2"
+    final static String SYMBOL_THREE_NAME = "BIO3"
     final static String BROKER_BASE_URL = "http://localhost:19082"
 
     def setupSpecWithSpring() {
@@ -59,114 +61,8 @@ class ILLBrokerSpec extends TestBase {
         return super.getBaseUrl()?.replace("null", "")
     }
 
-    Map sendXMLMessage(String url, String message, Map additionalHeaders, long timeout) {
-        Map result = [ messageStatus: EventISO18626IncomingAbstractService.STATUS_ERROR ]
 
-        HttpBuilder http_client = ApacheHttpBuilder.configure({
-            client.clientCustomizer({  HttpClientBuilder builder ->
-                   RequestConfig.Builder requestBuilder = RequestConfig.custom()
-                   requestBuilder.connectTimeout = timeout
-                   requestBuilder.connectionRequestTimeout = timeout
-                   requestBuilder.socketTimeout = timeout
-                   builder.defaultRequestConfig = requestBuilder.build()
-            })
-            request.uri = url
-            request.contentType = XML[0]
-            request.headers['accept'] = 'application/xml, text/xml'
-            additionalHeaders?.each{ k, v ->
-                request.headers[k] = v
-            }
-        })
 
-        def response = http_client.post {
-            request.body = message
-            response.failure({ FromServer fromServer ->
-                String errorMessage = "Error from address ${url}: ${fromServer.getStatusCode()} ${fromServer}"
-                log.error(errorMessage)
-                String responseStatus = fromServer.getStatusCode().toString() + " " + fromServer.getMessage()
-                throw new RuntimeException(errorMessage)
-            })
-            response.success({ FromServer fromServer, xml ->
-                String responseStatus = "${fromServer.getStatusCode()} ${fromServer.getMessage()}"
-                log.debug("Got response: ${responseStatus}")
-                if (xml != null) {
-                    result.rawData = groovy.xml.XmlUtil.serialize(xml)
-                } else {
-                    result.errorData = EventISO18626IncomingAbstractService.ERROR_TYPE_NO_XML_SUPPLIED
-                }
-
-            })
-        }
-        log.debug("Got response message: ${response}")
-
-        return result
-
-    }
-
-    private String waitForRequestState(String tenant, long timeout, String patron_reference, String required_state) {
-        Map params = [
-                'max':'100',
-                'offset':'0',
-                'match':'patronReference',
-                'term':patron_reference
-        ]
-        return waitForRequestStateParams(tenant, timeout, params, required_state)
-    }
-
-    private String waitForRequestStateById(String tenant, long timeout, String id, String required_state) {
-        Map params = [
-                'max':'1',
-                'offset':'0',
-                'match':'id',
-                'term':id
-        ]
-        return waitForRequestStateParams(tenant, timeout, params, required_state)
-    }
-
-    private String waitForRequestStateByHrid(String tenant, long timeout, String hrid, String required_state) {
-        Map params = [
-                'max':'1',
-                'offset':'0',
-                'match':'hrid',
-                'term':hrid
-        ]
-        return waitForRequestStateParams(tenant, timeout, params, required_state)
-    }
-
-    private String waitForRequestStateParams(String tenant, long timeout, Map params, String required_state) {
-        long start_time = System.currentTimeMillis()
-        String request_id = null
-        String request_state = null
-        long elapsed = 0
-        while ( ( required_state != request_state ) &&
-                ( elapsed < timeout ) ) {
-
-            setHeaders(['X-Okapi-Tenant': tenant])
-            // https://east-okapi.folio-dev.indexdata.com/rs/patronrequests?filters=isRequester%3D%3Dtrue&match=patronGivenName&perPage=100&sort=dateCreated%3Bdesc&stats=true&term=Michelle
-            def resp = doGet("${baseUrl}rs/patronrequests",
-                    params)
-            if (resp?.size() == 1) {
-                request_id = resp[0].id
-                request_state = resp[0].state?.code
-            } else {
-                log.debug("waitForRequestState: Request with params ${params} not found")
-            }
-
-            if (required_state != request_state) {
-                // Request not found OR not yet in required state
-                log.debug("Not yet found.. sleeping")
-                Thread.sleep(1000)
-            }
-            elapsed = System.currentTimeMillis() - start_time
-        }
-        log.debug("Found request on tenant ${tenant} with params ${params} in state ${request_state} after ${elapsed} milliseconds")
-
-        if ( required_state != request_state ) {
-            throw new Exception("Expected ${required_state} but timed out waiting, current state is ${request_state}")
-        }
-
-        return request_id
-    }
 
     private Map performActionFromFileAndCheckStatus(String performSupActionUrl, String actionFileName, String tenant,
             String reqStatus, String supStatus, String requesterTenantId, String responderTenantId,
@@ -193,7 +89,6 @@ class ILLBrokerSpec extends TestBase {
         Map data = doGet(patronRequestUrl) as Map;
         return data;
     }
-
 
     void "Attempt to delete any old tenants"(tenantid, name) {
         when:"We post a delete request"
