+++ service/src/integration-test/groovy/org/olf/TestBase.groovy
@@ -1,6 +1,14 @@
 package org.olf
 
-import org.olf.rs.GrailsEventIdentifier;
+import com.k_int.web.toolkit.refdata.RefdataCategory
+import com.k_int.web.toolkit.refdata.RefdataValue
+import groovyx.net.http.ApacheHttpBuilder
+import groovyx.net.http.FromServer
+import groovyx.net.http.HttpBuilder
+import org.apache.http.client.config.RequestConfig
+import org.apache.http.impl.client.HttpClientBuilder
+import org.olf.rs.GrailsEventIdentifier
+import org.olf.rs.statemodel.events.EventISO18626IncomingAbstractService;
 import org.springframework.beans.factory.annotation.Value;
 
 import com.k_int.web.toolkit.testing.HttpSpec;
@@ -9,7 +17,9 @@ import grails.events.bus.EventBus;
 import groovy.json.JsonBuilder;
 import groovy.util.logging.Slf4j;
 import spock.lang.Shared;
-import spock.util.concurrent.PollingConditions;
+import spock.util.concurrent.PollingConditions
+
+import static groovyx.net.http.ContentTypes.XML;
 
 @Slf4j
 class TestBase extends HttpSpec {
@@ -168,5 +178,139 @@ class TestBase extends HttpSpec {
 
         // Return the id of the location
         return(response?.id);
+    }
+
+    Map sendXMLMessage(String url, String message, Map additionalHeaders, long timeout) {
+        Map result = [ messageStatus: EventISO18626IncomingAbstractService.STATUS_ERROR ]
+
+        HttpBuilder http_client = ApacheHttpBuilder.configure({
+            client.clientCustomizer({ HttpClientBuilder builder ->
+                RequestConfig.Builder requestBuilder = RequestConfig.custom()
+                requestBuilder.connectTimeout = timeout
+                requestBuilder.connectionRequestTimeout = timeout
+                requestBuilder.socketTimeout = timeout
+                builder.defaultRequestConfig = requestBuilder.build()
+            })
+            request.uri = url
+            request.contentType = XML[0]
+            request.headers['accept'] = 'application/xml, text/xml'
+            additionalHeaders?.each{ k, v ->
+                request.headers[k] = v
+            }
+        })
+
+        def response = http_client.post {
+            request.body = message
+            response.failure({ FromServer fromServer ->
+                String errorMessage = "Error from address ${url}: ${fromServer.getStatusCode()} ${fromServer}"
+                log.error(errorMessage)
+                String responseStatus = fromServer.getStatusCode().toString() + " " + fromServer.getMessage()
+                throw new RuntimeException(errorMessage)
+            })
+            response.success({ FromServer fromServer, xml ->
+                String responseStatus = "${fromServer.getStatusCode()} ${fromServer.getMessage()}"
+                log.debug("Got response: ${responseStatus}")
+                if (xml != null) {
+                    result.rawData = groovy.xml.XmlUtil.serialize(xml)
+                } else {
+                    result.errorData = EventISO18626IncomingAbstractService.ERROR_TYPE_NO_XML_SUPPLIED
+                }
+
+            })
+        }
+        log.debug("Got response message: ${response}")
+
+        return result
+
+    }
+
+    String waitForRequestState(String tenant, long timeout, String patron_reference, String required_state) {
+        Map params = [
+                'max':'100',
+                'offset':'0',
+                'match':'patronReference',
+                'term':patron_reference
+        ]
+        return waitForRequestStateParams(tenant, timeout, params, required_state)
+    }
+
+    String waitForRequestStateById(String tenant, long timeout, String id, String required_state) {
+        Map params = [
+                'max':'1',
+                'offset':'0',
+                'match':'id',
+                'term':id
+        ]
+        return waitForRequestStateParams(tenant, timeout, params, required_state)
+    }
+
+    String waitForRequestStateByHrid(String tenant, long timeout, String hrid, String required_state) {
+        Map params = [
+                'max':'1',
+                'offset':'0',
+                'match':'hrid',
+                'term':hrid
+        ]
+        return waitForRequestStateParams(tenant, timeout, params, required_state)
+    }
+
+    String waitForRequestStateParams(String tenant, long timeout, Map params, String required_state) {
+        long start_time = System.currentTimeMillis()
+        String request_id = null
+        String request_state = null
+        long elapsed = 0
+        while ( ( required_state != request_state ) &&
+                ( elapsed < timeout ) ) {
+
+            setHeaders(['X-Okapi-Tenant': tenant])
+            // https://east-okapi.folio-dev.indexdata.com/rs/patronrequests?filters=isRequester%3D%3Dtrue&match=patronGivenName&perPage=100&sort=dateCreated%3Bdesc&stats=true&term=Michelle
+            def resp = doGet("${baseUrl}rs/patronrequests",
+                    params)
+            if (resp?.size() == 1) {
+                request_id = resp[0].id
+                request_state = resp[0].state?.code
+            } else {
+                log.debug("waitForRequestState: Request with params ${params} not found")
+            }
+
+            if (required_state != request_state) {
+                // Request not found OR not yet in required state
+                log.debug("Not yet found.. sleeping")
+                Thread.sleep(1000)
+            }
+            elapsed = System.currentTimeMillis() - start_time
+        }
+        log.debug("Found request on tenant ${tenant} with params ${params} in state ${request_state} after ${elapsed} milliseconds")
+
+        if ( required_state != request_state ) {
+            throw new Exception("Expected ${required_state} but timed out waiting, current state is ${request_state}")
+        }
+
+        return request_id
+    }
+
+    // For the given tenant fetch the specified request
+    Map fetchRequest(String tenant, String requestId) {
+
+        setHeaders([ 'X-Okapi-Tenant': tenant ]);
+        // https://east-okapi.folio-dev.indexdata.com/rs/patronrequests/{id}
+        def response = doGet("${baseUrl}rs/patronrequests/${requestId}")
+        return response;
+    }
+
+    protected String randomCrap(int length) {
+        String source = (('A'..'Z') + ('a'..'z')).join();
+        Random rand = new Random();
+        return (1..length).collect { source[rand.nextInt(source.length())]}.join();
+    }
+
+    // Rudely copy-pasted from EventMessageRequestIndService
+    RefdataValue findRefdataValue(String label, String vocabulary) {
+        RefdataCategory cat = RefdataCategory.findByDesc(vocabulary);
+        if (cat) {
+            RefdataValue rdv = RefdataValue.findByOwnerAndValue(cat, label);
+            return rdv;
+        }
+        return null;
     }
 }
